Presenting all products in categories fragment

+++ Amun-Store/app/src/main/java/com/example/amunstore/model/Options.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 
 data class Options (
 
-  @SerializedName("id"         ) var id        : Int?              = null,
-  @SerializedName("product_id" ) var productId : Int?              = null,
+  @SerializedName("id"         ) var id        : Long?              = null,
+  @SerializedName("product_id" ) var productId : Long?              = null,
   @SerializedName("name"       ) var name      : String?           = null,
   @SerializedName("position"   ) var position  : Int?              = null,
   @SerializedName("values"     ) var values    : ArrayList<String> = arrayListOf()
